# Added more optimized model for whisper small, Apply copy blob data API