--- a/audio_processing/whisper/whisper.py
+++ b/audio_processing/whisper/whisper.py
@@ -174,6 +174,7 @@ if not args.onnx:
     AILIA_VERSION_MINOR = int(version[1])
     AILIA_VERSION_REVISION = int(version[2])
     REQUIRE_CONSTANT_SHAPE_BETWEEN_INFERENCE = (AILIA_VERSION_MAJOR<=1 and AILIA_VERSION_MINOR<=2 and AILIA_VERSION_REVISION<14)
+    COPY_BLOB_DATA_ENABLE = (AILIA_VERSION_MAJOR<=1 and AILIA_VERSION_MINOR<=2 and AILIA_VERSION_REVISION>=15)
     SAVE_ENC_SHAPE = ()
     SAVE_DEC_SHAPE = ()
 
@@ -358,6 +359,7 @@ def get_audio_features(enc_net, mel):
 def inference_logits(dec_net, tokens, audio_features, kv_cache=None, initial_token_length=None, constant_audio_feature = False):
     n_group = tokens.shape[0]
     initial_token_length = initial_token_length if initial_token_length else tokens.shape[-1]
+    is_init_kv_cache = False
     if kv_cache is None:
         if not args.dynamic_kv_cache:
             kv_cache = new_kv_cache(n_group)
@@ -365,6 +367,7 @@ def inference_logits(dec_net, tokens, audio_features, kv_cache=None, initial_tok
             kv_cache = new_kv_cache(n_group, initial_token_length)
         offset = 0
         length = initial_token_length
+        is_init_kv_cache = True
     else:
         offset = kv_cache.shape[2]
         if not args.dynamic_kv_cache:
@@ -402,10 +405,18 @@ def inference_logits(dec_net, tokens, audio_features, kv_cache=None, initial_tok
 
             dec_net.predict([tokens, audio_features, kv_cache, offset], output = output)
         else:
-            if constant_audio_feature:
-                dec_net.predict({"tokens":tokens, "kv_cache":kv_cache, "offset":offset}, output = output)
+            if is_init_kv_cache or not COPY_BLOB_DATA_ENABLE:
+                if constant_audio_feature:
+                    dec_net.predict({"tokens":tokens, "kv_cache":kv_cache, "offset":offset}, output = output)
+                else:
+                    dec_net.predict([tokens, audio_features, kv_cache, offset], output = output)
             else:
-                dec_net.predict([tokens, audio_features, kv_cache, offset], output = output)
+                dec_net.copy_blob_data("kv_cache", "output_kv_cache", None)
+                output = [logits]
+                if constant_audio_feature:
+                    dec_net.predict({"tokens":tokens, "offset":offset}, output = output)
+                else:
+                    dec_net.predict({"tokens":tokens, "audio_features": audio_features, "offset":offset}, output = output)
     else:
         kv_cache = kv_cache.astype(np.float32)
         output = dec_net.run(None, {