--- a/audio_processing/whisper/whisper.py
+++ b/audio_processing/whisper/whisper.py
@@ -195,14 +195,14 @@ if args.normal:
 
 if not args.dynamic_kv_cache:
     # 高速化のためKV_CACHEのサイズを最大サイズで固定化したバージョン
-    WEIGHT_DEC_TINY_PATH = "decoder_tiny_fix_kv_cache"+ OPT +".onnx"
-    MODEL_DEC_TINY_PATH = "decoder_tiny_fix_kv_cache"+ OPT +".onnx.prototxt"
-    WEIGHT_DEC_BASE_PATH = "decoder_base_fix_kv_cache"+ OPT +".onnx"
-    MODEL_DEC_BASE_PATH = "decoder_base_fix_kv_cache"+ OPT +".onnx.prototxt"
+    WEIGHT_DEC_TINY_PATH = "decoder_tiny_fix_kv_cache"+ OPT2 +".onnx"
+    MODEL_DEC_TINY_PATH = "decoder_tiny_fix_kv_cache"+ OPT2 +".onnx.prototxt"
+    WEIGHT_DEC_BASE_PATH = "decoder_base_fix_kv_cache"+ OPT2 +".onnx"
+    MODEL_DEC_BASE_PATH = "decoder_base_fix_kv_cache"+ OPT2 +".onnx.prototxt"
     WEIGHT_DEC_SMALL_PATH = "decoder_small_fix_kv_cache"+ OPT2 +".onnx"
     MODEL_DEC_SMALL_PATH = "decoder_small_fix_kv_cache"+ OPT2 +".onnx.prototxt"
-    WEIGHT_DEC_MEDIUM_PATH = "decoder_medium_fix_kv_cache"+ OPT +".onnx"
-    MODEL_DEC_MEDIUM_PATH = "decoder_medium_fix_kv_cache"+ OPT +".onnx.prototxt"
+    WEIGHT_DEC_MEDIUM_PATH = "decoder_medium_fix_kv_cache"+ OPT2 +".onnx"
+    MODEL_DEC_MEDIUM_PATH = "decoder_medium_fix_kv_cache"+ OPT2 +".onnx.prototxt"
 else:
     # KV_CACHEが推論ごとに変化するバージョン
     WEIGHT_DEC_TINY_PATH = "decoder_tiny.onnx"