whisper.smallモデルをailia.optimizerで最適化。

||変更前|変更後|高速化比率|
|-----|-----|-----|-----|
|whisper small (M2 Mac CPU) |9266ms|7939ms|15%|
|whisper small (Core i7 CPU)|16610ms|13781ms|17%|
|whisper small (RTX3080 GPU)|7829ms|4141ms|47%|

また、cuDNN向けに、ailia SDK 1.2.15以降であればcopy_blob_data APIを使用するようにした。

opt

```
kyakuno@MacBook-Air whisper % python3 whisper.py -e 0 -b
 INFO arg_utils.py (13) : Start!
 INFO arg_utils.py (163) : env_id: 0
 INFO arg_utils.py (166) : CPU
 INFO model_utils.py (84) : ONNX file and Prototxt file are prepared!
 INFO model_utils.py (84) : ONNX file and Prototxt file are prepared!
 INFO whisper.py (903) : This model not optimized for macOS GPU currently. So we will use BLAS (env_id = 1).
 INFO whisper.py (807) : demo.wav
 INFO whisper.py (813) : Start inference...
 INFO whisper.py (815) : BENCHMARK mode
 INFO whisper.py (351) : 	encoder processing time 2547 ms
 INFO whisper.py (417) : 	decoder processing time 367 ms
 INFO whisper.py (677) : Detected language: English
 INFO whisper.py (351) : 	encoder processing time 2495 ms
 INFO whisper.py (417) : 	decoder processing time 400 ms
 INFO whisper.py (417) : 	decoder processing time 406 ms
 INFO whisper.py (417) : 	decoder processing time 400 ms
 INFO whisper.py (417) : 	decoder processing time 60 ms
 INFO whisper.py (417) : 	decoder processing time 55 ms
 INFO whisper.py (417) : 	decoder processing time 55 ms
 INFO whisper.py (417) : 	decoder processing time 53 ms
 INFO whisper.py (417) : 	decoder processing time 77 ms
 INFO whisper.py (417) : 	decoder processing time 67 ms
 INFO whisper.py (417) : 	decoder processing time 55 ms
 INFO whisper.py (417) : 	decoder processing time 55 ms
 INFO whisper.py (417) : 	decoder processing time 58 ms
 INFO whisper.py (417) : 	decoder processing time 56 ms
 INFO whisper.py (417) : 	decoder processing time 53 ms
 INFO whisper.py (417) : 	decoder processing time 53 ms
 INFO whisper.py (417) : 	decoder processing time 53 ms
 INFO whisper.py (417) : 	decoder processing time 54 ms
 INFO whisper.py (417) : 	decoder processing time 63 ms
 INFO whisper.py (417) : 	decoder processing time 56 ms
 INFO whisper.py (417) : 	decoder processing time 57 ms
 INFO whisper.py (417) : 	decoder processing time 53 ms
 INFO whisper.py (417) : 	decoder processing time 53 ms
 INFO whisper.py (417) : 	decoder processing time 53 ms
 INFO whisper.py (417) : 	decoder processing time 52 ms
 INFO whisper.py (417) : 	decoder processing time 53 ms
 INFO whisper.py (417) : 	decoder processing time 53 ms
 INFO whisper.py (417) : 	decoder processing time 53 ms
 INFO whisper.py (417) : 	decoder processing time 52 ms
 INFO whisper.py (417) : 	decoder processing time 53 ms
 INFO whisper.py (417) : 	decoder processing time 53 ms
 INFO whisper.py (417) : 	decoder processing time 53 ms
 INFO whisper.py (417) : 	decoder processing time 52 ms
 INFO whisper.py (417) : 	decoder processing time 53 ms
 INFO whisper.py (417) : 	decoder processing time 62 ms
 INFO whisper.py (417) : 	decoder processing time 110 ms
 INFO whisper.py (417) : 	decoder processing time 60 ms
 INFO whisper.py (417) : 	decoder processing time 55 ms
 INFO whisper.py (417) : 	decoder processing time 55 ms
 INFO whisper.py (417) : 	decoder processing time 55 ms
 INFO whisper.py (417) : 	decoder processing time 58 ms
 INFO whisper.py (417) : 	decoder processing time 57 ms
 INFO whisper.py (417) : 	decoder processing time 54 ms
 INFO whisper.py (417) : 	decoder processing time 55 ms
 INFO whisper.py (711) : [00:00.000 --> 00:10.000]  He hoped there would be stew for dinner, turnips and carrots and bruised potatoes and fat mutton pieces to be ladled out in thick, peppered, flour-fattened sauce.
 INFO whisper.py (821) : 	total processing time 9266 ms
 INFO whisper.py (830) : Script finished successfully.
```

opt2

```
kyakuno@MacBook-Air whisper % python3 whisper.py -e 0 -b
 INFO arg_utils.py (13) : Start!
 INFO arg_utils.py (163) : env_id: 0
 INFO arg_utils.py (166) : CPU
 INFO model_utils.py (84) : ONNX file and Prototxt file are prepared!
 INFO model_utils.py (69) : Downloading onnx file... (save path: decoder_small_fix_kv_cache.opt2.onnx)
 INFO model_utils.py (75) : ======================= 100.00% ( 756574KB )]

 INFO model_utils.py (77) : Downloading prototxt file... (save path: decoder_small_fix_kv_cache.opt2.onnx.prototxt)
 INFO model_utils.py (83) : ======================= 100.00% ( 487KB )]

 INFO model_utils.py (84) : ONNX file and Prototxt file are prepared!
 INFO whisper.py (905) : This model not optimized for macOS GPU currently. So we will use BLAS (env_id = 1).
 INFO whisper.py (809) : demo.wav
 INFO whisper.py (815) : Start inference...
 INFO whisper.py (817) : BENCHMARK mode
 INFO whisper.py (353) : 	encoder processing time 2264 ms
 INFO whisper.py (419) : 	decoder processing time 349 ms
 INFO whisper.py (679) : Detected language: English
 INFO whisper.py (353) : 	encoder processing time 2200 ms
 INFO whisper.py (419) : 	decoder processing time 330 ms
 INFO whisper.py (419) : 	decoder processing time 315 ms
 INFO whisper.py (419) : 	decoder processing time 351 ms
 INFO whisper.py (419) : 	decoder processing time 46 ms
 INFO whisper.py (419) : 	decoder processing time 45 ms
 INFO whisper.py (419) : 	decoder processing time 45 ms
 INFO whisper.py (419) : 	decoder processing time 45 ms
 INFO whisper.py (419) : 	decoder processing time 45 ms
 INFO whisper.py (419) : 	decoder processing time 57 ms
 INFO whisper.py (419) : 	decoder processing time 50 ms
 INFO whisper.py (419) : 	decoder processing time 47 ms
 INFO whisper.py (419) : 	decoder processing time 45 ms
 INFO whisper.py (419) : 	decoder processing time 44 ms
 INFO whisper.py (419) : 	decoder processing time 46 ms
 INFO whisper.py (419) : 	decoder processing time 46 ms
 INFO whisper.py (419) : 	decoder processing time 45 ms
 INFO whisper.py (419) : 	decoder processing time 45 ms
 INFO whisper.py (419) : 	decoder processing time 45 ms
 INFO whisper.py (419) : 	decoder processing time 45 ms
 INFO whisper.py (419) : 	decoder processing time 45 ms
 INFO whisper.py (419) : 	decoder processing time 45 ms
 INFO whisper.py (419) : 	decoder processing time 45 ms
 INFO whisper.py (419) : 	decoder processing time 45 ms
 INFO whisper.py (419) : 	decoder processing time 46 ms
 INFO whisper.py (419) : 	decoder processing time 45 ms
 INFO whisper.py (419) : 	decoder processing time 44 ms
 INFO whisper.py (419) : 	decoder processing time 45 ms
 INFO whisper.py (419) : 	decoder processing time 46 ms
 INFO whisper.py (419) : 	decoder processing time 45 ms
 INFO whisper.py (419) : 	decoder processing time 46 ms
 INFO whisper.py (419) : 	decoder processing time 45 ms
 INFO whisper.py (419) : 	decoder processing time 46 ms
 INFO whisper.py (419) : 	decoder processing time 46 ms
 INFO whisper.py (419) : 	decoder processing time 45 ms
 INFO whisper.py (419) : 	decoder processing time 45 ms
 INFO whisper.py (419) : 	decoder processing time 47 ms
 INFO whisper.py (419) : 	decoder processing time 46 ms
 INFO whisper.py (419) : 	decoder processing time 46 ms
 INFO whisper.py (419) : 	decoder processing time 46 ms
 INFO whisper.py (419) : 	decoder processing time 46 ms
 INFO whisper.py (419) : 	decoder processing time 45 ms
 INFO whisper.py (419) : 	decoder processing time 46 ms
 INFO whisper.py (419) : 	decoder processing time 46 ms
 INFO whisper.py (713) : [00:00.000 --> 00:10.000]  He hoped there would be stew for dinner, turnips and carrots and bruised potatoes and fat mutton pieces to be ladled out in thick, peppered, flour-fattened sauce.
 INFO whisper.py (823) : 	total processing time 7939 ms
 INFO whisper.py (832) : Script finished successfully.
```