@kyakuno 2023-07-24
whisper.tiny, base, mediumモデルも追加。