@kyakuno 2023-07-24
RTX3080 + cuDNN。

|モデル|変更前|変更後|
|-----|-----|-----|
|tiny|2313ms|2047ms|
|base|3031ms|2578ms|
|medium|12751ms|8734ms|